update caps for traveler
