+++ bin/src/main/resources/application.properties
@@ -1,5 +1,5 @@
-#spring.datasource.url=jdbc:mysql://utopiadb.cqohiovhe2ls.us-east-2.rds.amazonaws.com:3306/utopiadb?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-spring.datasource.url=jdbc:mysql://localhost:3306/utopiadb?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+spring.datasource.url=jdbc:mysql://utopiadb.cqohiovhe2ls.us-east-2.rds.amazonaws.com:3306/utopiadb?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
+#spring.datasource.url=jdbc:mysql://localhost:3306/utopiadb?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
 
 spring.datasource.username=root
 spring.datasource.password=root1234
